Allow for torrents of size greater than Int.MaxValue

+++ core/src/main/scala/com/cmhteixeira/bittorrent/swarm/TorrentFileManager.scala
@@ -6,12 +6,12 @@ import java.nio.file.Path
 import scala.concurrent.Future
 
 trait TorrentFileManager {
-  def write(file: Path, offset: Int, block: ByteVector): Future[Unit]
-  def read(file: Path, offset: Int, chunkSize: Int): Future[ByteVector]
+  def write(file: Path, offset: Long, block: ByteVector): Future[Unit]
+  def read(file: Path, offset: Long, chunkSize: Int): Future[ByteVector]
 
   def complete(slices: List[TorrentFileManager.FileSlice])(cond: ByteVector => Boolean): Future[Boolean]
 }
 
 object TorrentFileManager {
-  case class FileSlice(path: Path, offset: Int, size: Int)
+  case class FileSlice(path: Path, offset: Long, size: Int)
 }
